requirements.txt
pandas==2.2.2
python-dotenv==1.0.1
Requests==2.32.3
urllib3==2.2.2
